fix jerky zooming when close to end of scrollable area
The affected ScrollViewer's Offset is correctly bound to the TimelineState's scroll position, but whenever that offset was set it was being coerced to the maximum possible value for the current ScrollViewer's extent. However the extent was defined by the ScrollViewer's contents, which were updated afterwards, causing the offset being coerced to a too small value when zooming in.
The fix is to also bind the extent, so it is updated by the time the offset is being set.