fix jumpy scrolling and zooming
The mouse wheel delta is typically always 1 or -1, but if you scroll fast enough it can also be multiples of that.
I suppose that is to compensate for the mouse wheel scrolling faster that the mouse wheel events get processed.
The result feels jarring, so the fix is to just ignore the delta value.